add: 오류 처리 추가 및 수정

+++ src/main/java/com/solution/recipetalk/service/ingredient/trimming/row/impl/EditIngredientTrimmingRowServiceImpl.java
@@ -5,10 +5,18 @@ import com.solution.recipetalk.domain.ingredient.trimming.entity.IngredientTrimm
 import com.solution.recipetalk.domain.ingredient.trimming.repository.IngredientTrimmingRepository;
 import com.solution.recipetalk.domain.ingredient.trimming.row.entity.IngredientTrimmingRow;
 import com.solution.recipetalk.domain.ingredient.trimming.row.repository.IngredientTrimmingRowRepository;
+import com.solution.recipetalk.domain.user.entity.UserDetail;
+import com.solution.recipetalk.domain.user.repository.UserDetailRepository;
 import com.solution.recipetalk.dto.ingredient.trimming.row.IngredientTrimmingRowModifyDTO;
+import com.solution.recipetalk.exception.CustomException;
+import com.solution.recipetalk.exception.ErrorCode;
+import com.solution.recipetalk.exception.ingredient.trimming.IngredientTrimmingNotFoundException;
+import com.solution.recipetalk.exception.ingredient.trimming.row.IngredientTrimmingRowNotFoundException;
 import com.solution.recipetalk.exception.s3.ImageUploadFailedException;
+import com.solution.recipetalk.exception.user.UserNotFoundException;
 import com.solution.recipetalk.s3.upload.S3Uploader;
 import com.solution.recipetalk.service.ingredient.trimming.row.EditIngredientTrimmingRowService;
+import com.solution.recipetalk.util.ContextHolder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -24,13 +32,21 @@ public class EditIngredientTrimmingRowServiceImpl implements EditIngredientTrimm
     private final IngredientTrimmingRepository ingredientTrimmingRepository;
     private final IngredientTrimmingRowRepository ingredientTrimmingRowRepository;
     private final S3Uploader s3Uploader;
+    private final UserDetailRepository userDetailRepository;
 
     @Override
     public ResponseEntity<?> editIngredientTrimmingRow(List<IngredientTrimmingRowModifyDTO> dtoList, Long trimmingId) {
-        IngredientTrimming ingredientTrimming = ingredientTrimmingRepository.findById(trimmingId).orElseThrow();
+        Long userLoginId = ContextHolder.getUserLoginId();
+        UserDetail currentUser = userDetailRepository.findById(userLoginId).orElseThrow(UserNotFoundException::new);
+        IngredientTrimming ingredientTrimming = ingredientTrimmingRepository.findById(trimmingId).orElseThrow(IngredientTrimmingNotFoundException::new);
+
+        if (!ingredientTrimming.getBoard().getWriter().equals(currentUser)){
+            throw new CustomException(ErrorCode.NOT_AUTHORIZED_TO_MODIFY);
+        }
         List<IngredientTrimmingRow> ingredientTrimmingRows = dtoList.stream().map(dto -> {
             IngredientTrimmingRow ingredientTrimmingRow = ingredientTrimmingRowRepository
-                    .findByIngredientTrimmingAndTrimmingSeqAndTrimmingSubSeq(ingredientTrimming, dto.getTrimmingSeq(), dto.getTrimmingSubSeq()).orElseThrow();
+                    .findByIngredientTrimmingAndTrimmingSeqAndTrimmingSubSeq(ingredientTrimming, dto.getTrimmingSeq(), dto.getTrimmingSubSeq())
+                    .orElseThrow(IngredientTrimmingRowNotFoundException::new);
 
             s3Uploader.deleteFile(ingredientTrimmingRow.getImgURI(), S3dir.INGREDIENT_TRIMMING_ROW_IMG_DIR);
